complete overhaul to allow multiple years of AOC solutions to work in one repository.
This is done by globally setting a year (currently stored in `year.txt`). Each year of AOC solutions is its own package within the repository located at `year/`. and then filtering all commands (scaffold, download, try, solve, etc) to run in that package or only work within the package's files; most of that is done with the `src/` files, not the year's `src/` code. `year_template/` stores a template of a year's package.